requirements.txt
prefect>=2.0.0
python-chess>=1.999.0
google-cloud-bigquery>=3.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
python-dotenv>=0.20.0
requests>=2.26.0
db-dtypes>=1.0.0